+++ src/pages/index.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react'
 
 import Layout from '@/components/layout/Layout'
+import CV from '@/components/MainPage/CV'
 import Introductory from '@/components/MainPage/Introductory'
 import Seo from '@/components/Seo'
 
@@ -8,8 +9,9 @@ export default function LearnPage(): JSX.Element {
 	return (
 		<Layout>
 			<Seo />
-			<div className='container py-2'>
+			<div className='container py-2 md:px-36 dark:text-white'>
 				<Introductory />
+				<CV />
 			</div>
 		</Layout>
 	)