+++ tailwind.config.js
@@ -70,8 +70,7 @@ module.exports = {
 				flicker: {
 					'0%, 19.999%, 22%, 62.999%, 64%, 64.999%, 70%, 100%': {
 						opacity: 0.99,
-						filter:
-							'drop-shadow(0 0 1px rgba(252, 211, 77)) drop-shadow(0 0 15px rgba(245, 158, 11)) drop-shadow(0 0 1px rgba(252, 211, 77))',
+						filter: 'drop-shadow(0 0 1px rgba(252, 211, 77)) drop-shadow(0 0 15px rgba(245, 158, 11)) drop-shadow(0 0 1px rgba(252, 211, 77))',
 					},
 					'20%, 21.999%, 63%, 63.999%, 65%, 69.999%': {
 						opacity: 0.4,